pyproject.toml
[tool.poetry]
name = "chord-liner"
version = "0.1.0"
description = ""
authors = ["isudzumi <isudzumi@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
chalice = "^1.12.0"
line-bot-sdk = "^1.15.0"
pychord = "^0.4.1"
numpy = "^1.18.1"

[tool.poetry.dev-dependencies]
pylint = "^2.4.4"
mypy = "^0.761"
rope = "^0.16.0"
boto3 = "^1.11.9"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
